# [libc][math] Improve the performance of sqrtf128.