--- a/libc/src/math/generic/sqrtf128.cpp
+++ b/libc/src/math/generic/sqrtf128.cpp
@@ -19,6 +19,9 @@
 // arithmetic by Alexei Sibidanov (sibid@uvic.ca):
 //   https://github.com/sibidanov/llvm-project/tree/as_sqrt_v2
 //   https://github.com/sibidanov/llvm-project/tree/as_sqrt_v3
+// TODO: Update the reference once Alexei's implementation is in the CORE-MATH
+// project. https://github.com/llvm/llvm-project/issues/126794
+
 // Let the input be expressed as x = 2^e * m_x,
 // - Step 1: Range reduction
 //   Let x_reduced = 2^(e % 2) * m_x,