… libc/src/math/generic/CMakeLists.txt
… libc/test/UnitTest/FPMatcher.h
… libc/test/src/math/performance_testing/CMakeLists.txt
… libc/test/src/math/performance_testing/sqrtf128_perf.cpp
… libc/test/src/math/smoke/generic_sqrt_test.cpp
… libc/test/src/math/smoke/generic_sqrtf128_test.cpp
… libc/test/src/math/smoke/generic_sqrtf_test.cpp
… libc/test/src/math/smoke/generic_sqrtl_test.cpp
… libc/test/src/math/smoke/sqrt_test.cpp
… libc/test/src/math/smoke/sqrtf128_test.cpp
… libc/test/src/math/smoke/sqrtf16_test.cpp
… libc/test/src/math/smoke/sqrtf_test.cpp
… libc/test/src/math/smoke/sqrtl_test.cpp